RFI ALT/requirements.txt
pandas
openpyxl
pdfplumber
tqdm
python-dotenv
office365-rest-python-client